requirements.txt

### requirements.txt

```plaintext
tweepy==4.3.0
python-dotenv==0.19.0
